Added Player Disconnect Functionality
WinLossPopupView now references Gomoku in order to get the ip address
(ip was hardcoded before).
New method in GameViewController, showPlayerDiscon() which is used for
showing if an opponent has disconnected.
NetworkedPlayer has a new boolean at the top named winConMet, and sets
true if a win condition is met, that way player disconnect doesn't show
up when the game ends.
Finally, changed some of the variable names for consistency.